added modal for error file form
